refactor testing into different python files for tool_manager and import cleanup
